Fix working PoC

+++ src/rule.rs
@@ -1,14 +1,14 @@
 #[derive(Debug, Clone)]
 pub struct Rules {
-    min_age_hours: u64,
+    min_age_hours: u32,
 }
 
 impl Rules {
-    pub fn new(min_age_hours: u64) -> Self {
+    pub fn new(min_age_hours: u32) -> Self {
         Rules { min_age_hours }
     }
 
-    pub fn min_age_hours(&self) -> u64 {
+    pub fn min_age_hours(&self) -> u32 {
         self.min_age_hours
     }
 }
